refactor(filter-shrinkwrap): divide into several files
