+++ packages/filter-shrinkwrap/src/normalizeShrinkwrap.ts
@@ -1,0 +1,8 @@
+import { Shrinkwrap } from '@pnpm/shrinkwrap-types'
+
+export default function normalizeShrinkwrap (shr: Shrinkwrap) {
+  if (!shr.registry) {
+    delete shr.registry
+  }
+  return shr
+}
